pyproject.toml
[project]
name = "greenpowersweden"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiohappyeyeballs==2.6.1",
    "aiohttp==3.13.2",
    "aiosignal==1.4.0",
    "altair==5.5.0",
    "annotated-types==0.7.0",
    "anyio==4.11.0",
    "argon2-cffi==25.1.0",
    "argon2-cffi-bindings==25.1.0",
    "arrow==1.4.0",
    "asttokens==3.0.0",
    "async-lru==2.0.5",
    "attrs==25.4.0",
    "babel==2.17.0",
    "backoff==2.2.1",
    "bcrypt==5.0.0",
    "beautifulsoup4==4.14.2",
    "bleach==6.3.0",
    "blinker==1.9.0",
    "build==1.3.0",
    "cachetools==6.2.2",
    "certifi==2025.11.12",
    "cffi==2.0.0",
    "charset-normalizer==3.4.4",
    "chromadb==1.3.5",
    "click==8.3.1",
    "colorama==0.4.6",
    "coloredlogs==15.0.1",
    "comm==0.2.3",
    "cryptography==46.0.3",
    "dataclasses-json==0.6.7",
    "debugpy==1.8.17",
    "decorator==5.2.1",
    "defusedxml==0.7.1",
    "distro==1.9.0",
    "durationpy==0.10",
    "et-xmlfile==2.0.0",
    "executing==2.2.1",
    "fastjsonschema==2.21.2",
    "filelock==3.20.0",
    "filetype==1.2.0",
    "flatbuffers==25.9.23",
    "fqdn==1.5.1",
    "frozenlist==1.8.0",
    "fsspec==2025.10.0",
    "gdown==4.7.3",
    "gitdb==4.0.12",
    "gitpython==3.1.45",
    "google-ai-generativelanguage==0.9.0",
    "google-api-core==2.28.1",
    "google-auth==2.43.0",
    "googleapis-common-protos==1.72.0",
    "greenlet==3.2.4",
    "grpcio==1.76.0",
    "grpcio-status==1.76.0",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httptools==0.7.1",
    "httpx==0.28.1",
    "httpx-sse==0.4.3",
    "huggingface-hub==0.36.0",
    "humanfriendly==10.0",
    "idna==3.11",
    "importlib-metadata==8.7.0",
    "importlib-resources==6.5.2",
    "ipykernel==7.1.0",
    "ipython==9.7.0",
    "ipython-pygments-lexers==1.1.1",
    "ipywidgets==8.1.8",
    "isoduration==20.11.0",
    "jedi==0.19.2",
    "jinja2==3.1.6",
    "joblib==1.5.2",
    "json5==0.12.1",
    "jsonpatch==1.33",
    "jsonpointer==3.0.0",
    "jsonschema==4.25.1",
    "jsonschema-specifications==2025.9.1",
    "jupyter-client==8.6.3",
    "jupyter-core==5.9.1",
    "jupyter-events==0.12.0",
    "jupyter-lsp==2.3.0",
    "jupyter-server==2.17.0",
    "jupyter-server-terminals==0.5.3",
    "jupyterlab==4.4.10",
    "jupyterlab-pygments==0.3.0",
    "jupyterlab-server==2.28.0",
    "jupyterlab-widgets==3.0.16",
    "kubernetes==34.1.0",
    "langchain==1.1.0",
    "langchain-chroma==1.0.0",
    "langchain-classic==1.0.0",
    "langchain-community==0.4.1",
    "langchain-core==1.1.0",
    "langchain-google-genai==3.2.0",
    "langchain-huggingface==1.1.0",
    "langchain-text-splitters==1.0.0",
    "langgraph==1.0.3",
    "langgraph-checkpoint==3.0.1",
    "langgraph-prebuilt==1.0.4",
    "langgraph-sdk==0.2.9",
    "langsmith==0.4.43",
    "lark==1.3.1",
    "lxml==6.0.2",
    "markdown-it-py==4.0.0",
    "markupsafe==3.0.3",
    "marshmallow==3.26.1",
    "matplotlib-inline==0.2.1",
    "mdurl==0.1.2",
    "mistune==3.1.4",
    "mmh3==5.2.0",
    "mpmath==1.3.0",
    "multidict==6.7.0",
    "mypy-extensions==1.1.0",
    "narwhals==2.12.0",
    "nbclient==0.10.2",
    "nbconvert==7.16.6",
    "nbformat==5.10.4",
    "nest-asyncio==1.6.0",
    "networkx==3.5",
    "notebook==7.4.7",
    "notebook-shim==0.2.4",
    "numpy==2.3.4",
    "oauthlib==3.3.1",
    "onnxruntime==1.23.2",
    "openpyxl==3.1.5",
    "opentelemetry-api==1.38.0",
    "opentelemetry-exporter-otlp-proto-common==1.38.0",
    "opentelemetry-exporter-otlp-proto-grpc==1.38.0",
    "opentelemetry-proto==1.38.0",
    "opentelemetry-sdk==1.38.0",
    "opentelemetry-semantic-conventions==0.59b0",
    "orjson==3.11.4",
    "ormsgpack==1.12.0",
    "overrides==7.7.0",
    "packaging==25.0",
    "pandas==2.3.3",
    "pandocfilters==1.5.1",
    "parso==0.8.5",
    "pdf2image==1.17.0",
    "pdfminer-six==20251107",
    "pdfplumber==0.11.8",
    "pillow==12.0.0",
    "platformdirs==4.5.0",
    "posthog==5.4.0",
    "prometheus-client==0.23.1",
    "prompt-toolkit==3.0.52",
    "propcache==0.4.1",
    "proto-plus==1.26.1",
    "protobuf==6.33.1",
    "psutil==7.1.3",
    "pure-eval==0.2.3",
    "pyarrow==21.0.0",
    "pyasn1==0.6.1",
    "pyasn1-modules==0.4.2",
    "pybase64==1.4.2",
    "pycparser==2.23",
    "pydantic==2.12.4",
    "pydantic-core==2.41.5",
    "pydantic-settings==2.12.0",
    "pydeck==0.9.1",
    "pygments==2.19.2",
    "pypdfium2==5.0.0",
    "pypika==0.48.9",
    "pyproject-hooks==1.2.0",
    "pytesseract==0.3.13",
    "python-dateutil==2.9.0.post0",
    "python-docx==1.2.0",
    "python-dotenv==1.2.1",
    "python-json-logger==4.0.0",
    "pytz==2025.2",
    "pyyaml==6.0.3",
    "pyzmq==27.1.0",
    "referencing==0.37.0",
    "regex==2025.11.3",
    "requests==2.32.5",
    "requests-oauthlib==2.0.0",
    "requests-toolbelt==1.0.0",
    "rfc3339-validator==0.1.4",
    "rfc3986-validator==0.1.1",
    "rfc3987-syntax==1.1.0",
    "rich==14.2.0",
    "rpds-py==0.28.0",
    "rsa==4.9.1",
    "safetensors==0.6.2",
    "scikit-learn==1.7.2",
    "scipy==1.16.3",
    "send2trash==1.8.3",
    "sentence-transformers==5.1.2",
    "setuptools==80.9.0",
    "shellingham==1.5.4",
    "six==1.17.0",
    "smmap==5.0.2",
    "sniffio==1.3.1",
    "soupsieve==2.8",
    "sqlalchemy==2.0.44",
    "stack-data==0.6.3",
    "streamlit==1.51.0",
    "streamlit-pdf-viewer==0.0.26",
    "sympy==1.14.0",
    "tenacity==9.1.2",
    "terminado==0.18.1",
    "threadpoolctl==3.6.0",
    "tinycss2==1.4.0",
    "tokenizers==0.22.1",
    "toml==0.10.2",
    "torch==2.9.1",
    "tornado==6.5.2",
    "tqdm==4.67.1",
    "traitlets==5.14.3",
    "transformers==4.57.1",
    "typer==0.20.0",
    "typing-extensions==4.15.0",
    "typing-inspect==0.9.0",
    "typing-inspection==0.4.2",
    "tzdata==2025.2",
    "uri-template==1.3.0",
    "urllib3==2.3.0",
    "uvicorn==0.38.0",
    "watchdog==6.0.0",
    "watchfiles==1.1.1",
    "wcwidth==0.2.14",
    "webcolors==25.10.0",
    "webencodings==0.5.1",
    "websocket-client==1.9.0",
    "websockets==15.0.1",
    "widgetsnbextension==4.0.15",
    "xxhash==3.6.0",
    "yarl==1.22.0",
    "zipp==3.23.0",
    "zstandard==0.25.0",
]
